+++ src/component/NewPost/index.jsx
